+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog for lambda
 
+- [15.11.2020]
+    - add `-Wall` flag and fixes all messages
+    - refine parsing of lists such that the item type can be specified.
+      this change was required due to ambigious resolution of
+      ad-hoc polymorphism in case of empty lists.
 - [14.11.2020] initial commit
 
 ## Unreleased changes